+++ asab/metrics/influxdb.py
@@ -95,7 +95,7 @@ class InfluxDBTarget(asab.Configurable):
 		self.WriteURL = "{}{}".format(self.BaseURL, self.WriteRequest)
 
 		# Proactor service is used for alternative delivery of the metrics into the InfluxDB
-		# It is handly when a main loop can become very busy
+		# It is handy when a main loop can become very busy
 		if self.Config.getboolean('proactor'):
 			try:
 				from ..proactor import Module
@@ -137,23 +137,24 @@ class InfluxDBTarget(asab.Configurable):
 
 		try:
 			conn.request("POST", self.WriteRequest, rb, self.Headers)
+			response = conn.getresponse()
+			if response.status != 204:
+				L.warning("Failed to send metrics to InfluxDB.", struct_data={"url": self.BaseURL, "response.status": response.status, "response": response.read().decode("utf-8")})
+		except http.client.RemoteDisconnected:
+			L.error("Failed to send metrics to InfluxDB: Remote end closed connection without response.", struct_data={"url": self.BaseURL})
+			return
 		except (ConnectionError, socket.gaierror):
 			L.error("Failed to connect to InfluxDB.", struct_data={"url": self.BaseURL})
 			return
 		except Exception as err:
 			L.exception("Failed to send metrics to InfluxDB: {}".format(err), struct_data={"url": self.BaseURL})
 			return
-
-		response = conn.getresponse()
-		if response.status != 204:
-			L.warning(
-				"Sending metrics to InfluxDB failed.",
-				struct_data={
-					"url": self.BaseURL,
-					"response.status": response.status,
-					"response": response.read().decode("utf-8")
-				}
-			)
+		finally:
+			try:
+				conn.close()
+			except Exception as e:
+				# TODO: If this is too noisy, we can make it a debug log instead of warning
+				L.warning("Exception while closing InfluxDB connection: {}".format(e))
 
 
 def get_field(fk, fv):